// api/src/modules/transfers/api/models/aCHBatchTransferDto.ts
/**
 * Transfers API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ACHPaymentDto } from './aCHPaymentDto';

export interface ACHBatchTransferDto {
  /**
   * The id of batch payment
   */
  batchPaymentId: string;
  /**
   * Additional data represented as key-value pairs.
   */
  metadata?: any;
  /**
   * The url of webhook where the payment status will be send
   */
  webhookClientURL?: string;
  /**
   * Either credit or debit.
   */
  transaction: ACHBatchTransferDtoTransaction;
  /**
   * List of ACH Payments
   */
  payments: Array<ACHPaymentDto>;
  /**
   * Either normal or high. For ACH payments, high represents a same-day ACH
   */
  priority?: ACHBatchTransferDtoPriority;

  transferType?: 'ach';
}
export enum ACHBatchTransferDtoTransaction {
  Credit = 'credit',
  Debit = 'debit',
}
export enum ACHBatchTransferDtoPriority {
  Normal = 'normal',
  High = 'high',
}