// api/src/modules/transfers/api/models/expediteTransferDto.ts
/**
 * Transfers API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RtpPaymentDto } from './rtpPaymentDto';
import { TransferType } from './transfer-type';

export interface ExpediteTransferDto {
  /**
   * The id of batch payment
   */
  batchPaymentId: string;
  /**
   * Additional data represented as key-value pairs.
   */
  metadata?: any;
  /**
   * The url of webhook where the payment status will be send
   */
  webhookClientURL?: string;
  transferType: TransferType;
  /**
   * List of RTP Payments
   */
  payments: Array<RtpPaymentDto>;
}